Fix a bug, left shift overflow

+++ src/kernel/interrupt.c
@@ -203,7 +203,7 @@ void idt_init()
     pic_init();                              // 初始化8259A
 
     // 加载IDT
-    uint64_t idt_operand = (sizeof(idt) - 1) | ((uint64_t)((uint32_t)idt << 16));
+    uint64_t idt_operand = (sizeof(idt) - 1) | ((uint64_t)idt << 16);
     asm volatile("lidt %0" : : "m" (idt_operand));
     put_str("idt_init done\n");
 }
